backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.26.0
PyMuPDF==1.23.21
pydub==0.25.1
gTTS==2.5.0
pydantic==2.6.1
python-dotenv==1.0.1
aiofiles==23.2.1